+++ odoo/local-src/wim_queue/models/queue_job_log.py
@@ -7,14 +7,11 @@ RESULT_STATE = [("success", "Successful"), ("failed", "Failed")]
 
 
 class QueueJobLog(models.Model):
-
     _name = "queue.job.log"
     _description = "Log queue job requests"
 
     queue_job_id = fields.Many2one("queue.job", index=True, ondelete='cascade')
-    job_state = fields.Selection(
-        string="Job State", related="queue_job_id.state"
-    )
+    job_state = fields.Selection(string="Job State", related="queue_job_id.state")
     state = fields.Selection(
         RESULT_STATE,
         string="Action State",